Home page implemented
