+++ Xcode/Cardier/view/page/my-page/MyPageView.swift
@@ -1,0 +1,18 @@
+//
+//  MyPageView.swift
+//  Cardier
+//
+//  Created by John Choi on 1/26/24.
+//
+
+import SwiftUI
+
+struct MyPageView: View {
+    var body: some View {
+        Text("This is My Page")
+    }
+}
+
+#Preview {
+    MyPageView()
+}
